+++ __tests__/fixtures/test-v11-prototype/app/assets/javascripts/application.js
@@ -1,4 +1,5 @@
 /* global $ */
+/* global GOVUK */
 
 // Warn about using the kit in production
 if (window.console && window.console.info) {
@@ -7,4 +8,16 @@ if (window.console && window.console.info) {
 
 $(document).ready(function () {
   window.GOVUKFrontend.initAll()
+  // Use GOV.UK shim-links-with-button-role.js to trigger a link styled to look like a button,
+  // with role="button" when the space key is pressed.
+  GOVUK.shimLinksWithButtonRole.init()
+
+  // Details/summary polyfill from frontend toolkit
+  GOVUK.details.init()
+
+  // Show and hide toggled content
+  // Where .multiple-choice uses the data-target attribute
+  // to toggle hidden content
+  var showHideContent = new GOVUK.ShowHideContent()
+  showHideContent.init()
 })